Add files via upload
The task is to write a program which finds line segments on black and white images. The lines will always be one pixel wide, black on a white background, no grayscale or other colors. Each image will contain up to 2 lines. The output of your program should be the coordinates of the start and end of each line segment if there are any. The coordinates will be always whole pixel coordinates.
The images will be represented by text files. One row in the file is one row of the image. Each character in the row is one pixel. A pixel can have a value of 0 or 1. 0 means, this pixel is not part of a line (it’s white). 1 means, this pixel is part of a line (it’s black).
As usual, when representing images, the top-left pixel has (0, 0) coordinates. The X axis increases to the right, whereas the Y axis increases downwards.
In the regular input set the lines will not be touching or intersecting each other. However, in the for-extra-credit input set there can be touching or intersecting lines.
The output of your program should have one row for each image. The row should begin with the filename of the image, followed by the coordinates of the endpoints. Each coordinate should be surrounded by parentheses. The first value for each coordinate should be the X position, the second the Y position. The values should be separated by comas. The coordinates should be separated by spaces. The two endpoints of the same line should follow each other in the output.